requirements.txt
streamlit
python-dotenv
cohere
trulens_eval==0.32.0  # Ensure this version is correct
Huggingface  # Adjust version as needed
